Relocate rockets slice file
